+++ railhook-ui/src/types/api.generated.ts
@@ -6249,6 +6249,7 @@ export interface components {
             id?: string;
             /** Format: uuid */
             eventId?: string;
+            eventType?: string;
             /** Format: uuid */
             endpointId?: string;
             /** Format: uuid */